autobot_env/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
python-dotenv==1.0.1
langchain-openai==0.1.21
langgraph==0.2.60
langgraph-checkpoint>=2.1.1
